nouveau: manage nouveau_drm lifetime with devres
Make use of the devm_drm_dev_alloc() API to tie the lifetime of
nouveau_drm structure lives as long as the drm_device. This is important
because a reference to nouveau_drm is accessible from drm_device and is
used in, for example, the postclose callbacks which can occur after the
release callback is run.

Signed-off-by: Jeremy Cline <jcline@redhat.com>