nouveau: add a dedicated mutex for the client list in nouveau_drm
The list is currently protected by a mutex in nouveau_drm.client, but
this client may be deallocated prior to the final file closure. Add a
new lock and clean up the client list.

TODO: this is awful hacky code, mostly just to test if it fixes the
use-after-free. See what can be bound to the drm_device deallocation and
if we can actually just get rid of the list which appears unused.

Signed-off-by: Jeremy Cline <jcline@redhat.com>